[Test][E2E] Address review feedback: de-duplicate rescale ITs, add Javadoc, use RestConstant
